Make widgetsexplorer almost build
Lots of slaughtering going on here, as we're ripping out
QGraphics*-based UIs. The widgetExplorer object won't have UI in it
anymore, as that is done in the desktop shell package (rather than its
own package).

Lots of features are disabled right now (install, uninstall, download),
the idea is to first just get the model working, and then re-enable
features one at a time.