Use QScreen::virtualGeometry() for computing panel struts
QScreen::availableGeometry() returns screen size minus the struts, which is fine on first Plasma start. But if Plasma crashes, the struts set on the startup are not cleared and then when Plasma autorestarts, availableGeometry() gives it rect with the old struts to which Plasma adds one more panel height (so the total strut height is panel.height()*(numberOfCrashes+1)).

Using QScreen::virtualGeometry() gives us the whole screen including
struts and the new struts are computed and set properly.

Reviewed-by: Marco Martin