[kglobalaccel] Remove notification support
KGlobalAccel emitted notifications when:
* a shortcut is pressed
* a new shortcut is registered

Both are configured with no action at all. Thus the notification is not
of much use. Why it shouldn't show a popup had been discussed on kcd [1].

This means the notification right now has nothing more than debug
purpose. While this might be a valid usecase it doesn't make much sense
to do this with KNotification - for this see Aaron's mail [2]. Also e.g.
KWin dropped all notifications for debug purposes for the same reason.

If there is a need for a kind of notification on global shortcut
triggered or a new registered global shortcut this could also be easily
emulated by adding an explicit signal to the DBus interface.

This removes the KNotificiation dependency.

REVIEW: 117464

[1] http://lists.kde.org/?t=126463249400002&r=1&w=2&n=16
[2] http://lists.kde.org/?l=kde-core-devel&m=126463340225306&w=2