React to containment status changes
At the moment the Panel didn't have any code to react status changes from
the containment (and therefore its applets).

This patch aims to add this, only problem being that it doesn't work. The
"unhide requested" and "unhiding" messages are being displayed though.

I've been looking for the code that does the actual display of the
auto-hiding panels and I wasn't able to find it, help is welcome.

REVIEW: 117563