Improve wallpapers deduplication code a bit
On my system, I have a symlink /usr/share/wallpapers/kde-default.png ->
/usr/share/wallpapers/Elarun/contents/images/someresolution.png. This
symlink gets processed as an additional image even though it actually
points to an image which is already in the set.

So I changed it to always work with the target rather than symlink and
then it also checks for the target possibly being in a package, if it
is, then it processes the whole package rather than only one file from
it.

This fixes one wallpaper (same path) to be shown twice in the wallpaper
select dialog.

REVIEW: 117601