[kglobalaccel] Update X11 appTime from key press events
KGlobalAccelD sends the current appTime through DBus to the application
and KF5::GlobalAccel updates the appTime from the timestamp. This is
needed for following actions to succeed. E.g. KWin needs the updated
appTime for grabbing the keyboard following the kill window shortcut.

To get the current timestamp we just use the timestamp of the key press
event delivered to kglobalacceld. These events are filtered out, so that
Qt doesn't see them. Thus we need to update the timestamp and cannot hope
for Qt to update them.

REVIEW: 117839