Add unit test for kscreenlocker_greet to test all signals
The idea behind this unit test is to verify that no matter which signal
is sent to the greeter it doesn't return an exit code of 0 which would
indicate a successful authentication.

The test starts the greeter (needs to be installed) and waits 5 sec to
make sure that the greeter is really up and running. If the signal is
sent to early it's causing a CrashExit, which is fine but not the
expected result. There's an additional test to verify that the early sent
signal results in CrashExit.

REVIEW: 117893