Fix memory leaking in notifications
Up to this patch there was a new Dialog created for each and every new
notification incoming. After the notification was timed out, it would
close and delete the Dialog, but somehow that didn't free all the memory
and over time plasma-shell could get huuuge mem usage.

Now the notifications reuse only 3 Dialogs the whole lifetime --> memory
does not skyrockets anymore.

Note that the original leak in Dialog might still be present, but
notifications do not expose it anymore.

BUG: 333029
REVIEW: 117903