Actions to add apps to the panel/desktop containments and the task manager.
Containment mutability is taken into account, the smart thing
is done for Folder View (symlink the .desktop file, don't spawn
an Icon applet) and the Task Manager (if found) is asked if it
already has a launcher.

We're pretty unhappy with all the coupling, though it wasn't any
better in Plasma 1. Long-term, a more generic framework to re-
gister launchers and allow containments and applets to work with
that data is to be implemented.