Do not set a custom format on QQuickWindows
All QQuickWindows share a default format. By setting a custom format
it overwrites the default and this results in a format without a
stencil buffer being requested resulting in possible rendering errors.

This was used to get an alpha buffer. Instead use the static method
on QQuickWindow to request an alpha buffer.

REVIEW: 118123