[klipper] Show filter widget before sending key events to it
Klipper is supposed to show a filter widget when starting to type.
This is done by forwarding the QKeyEvent to the not visible widget.
This seems to have broken in the Qt5 port. Instead of passing to the
widget it's send again to the KlipperPopup resulting in recursive calls
to the same method which in the end crashes Klipper.

By showing the widget before sending it the key event, everything is
fine. In case that the widget is empty after processing the key event
it gets hidden nevertheless.

REVIEW: 118204