[screenlocker] Pass grace time as command line argument to greeter
Instead of having the greeter read the settings of the KSLD, KSLD passes
the required values to the greeter as command line arguments. This
ensures that the greeter doesn't show a lock UI while the KSLD is still
in grace time and vice versa.

Two command line options are added:
--graceTime <milliseconds>
--nolock

The option --graceTime starts the timer to show the lock ui on timeout.
The option --nolock is for the case that the KSLD never locks the
screen.

REVIEW: 118214