kscreenlocker_greet: use SA_RESTART for signal handler
As discussed in https://git.reviewboard.kde.org/r/117091/. Not using the
SA_RESTART flag might (in theory) cause the greeter to be aborted
(because certain syscalls may be interrupted and fail with EINTR).
SA_RESTART seems to be the BSD default and is used by "legacy" signal()
by default in glibc 2 and later as well anyway.

REVIEW: 118563