Port libtaskmanager away from QDesktopWidget
plasmoid.screen doesn't map to QDesktopWidget indexes anymore, therefore we
need to port it.
This patch uses the screen geometry to figure out what's the screen and then
passes around the screen rect so that we can filter out the screens that aren't
inside if the user asks for it.

REVIEW: 118548