X11/XCB becomes a required component
So far plasma-desktop supported to be built without X11 support (Win32),
but that was completely broken (X11Extras was required). Based on
discussion in [1] X11/XCB becomes a required component.

This means that we should not have any components in the build system
which are built optionaly depending on whether X11 is available. Ifdefs
in the code should stay, though.

For the future we will turn X11 in a optional dependency again. But it
will never be that a component won't be built if we are building without
X11. Instead there will be a required Wayland backend and this means all
X11 specific code needs to become platform aware.

So the way forward is:
1. X11 required
2. X11 required, Wayland optional
3. X11 required, Wayland required
4. Wayland required, X11 optional

REVIEW: 118694

[1] https://mail.kde.org/pipermail/plasma-devel/2014-June/032259.html