Always show battery remaining time
I am really tired of this topic. Yes, the battery remaining time is not that accurate and
jumps around (the latter is something that could be fixed using a moving average). But if
we jump onto that train we would also need to remove the battery percentage itself as it's not
accurate either. 30% on a dead battery can quickly become 5% and then the notebook suddenly
turns off. "Battery: Present".

The remaining time is in the battery's tooltip, many distributions have even patched the
battery monitor to show the time by default, there's even a clone of the original battery
monitor on GHNS just with that option enabled; every other desktop environment even emphasizes
the remaining time, Windows prefers the time over percentage, Mac OS X allows you to put
the remaining time in the panel rather than the percentage, etc etc.

As a maintainer I now made the decision to show remaining time by default - it's no
"in your face" display but in the battery's detail section - and I would kindly
ask all the people involved in this discussion to respect this. Thank you!

BUG: 290578