[klipper] Properly reset History::m_topIsUserSelected
When inserting an existing clipboard item again it became marked as a
user selected top because the call sequence to set m_topIsUserSelected
to false was incorrect.

Also when removing a user selected item it didn't get reset causing the
next (not selected) item to be marked as user selected.