Make sure we get to decide where our panels go
Workaround a lack of API from Qt where it will decide to move the panels
around without us being able to control it.
This patch overrides the QWindow::screenDestroyed slot to be able to decide
where to move the Panel or whether it needs to be removed. This is
considered a workaround and should be removed once the new API is in and
released.

See also: https://codereview.qt-project.org/#/c/88351/
BUG: 335710