Revert "set QSG_RENDER_LOOP as basic if not explicitly set"
This reverts commit 30d2489930ac8899e4e2a1e3d65312ad1c3ffa86.

The test results we got indicate that on NVIDIA the problems with the
threaded rendering loop do not exist and at the same time the rendering
result improves a lot. As Qt still disables the threaded rendering loop
for the problematic drivers, it's better to not force the basic rendering
loop.