Move path argument into KIconLoader ctor
This is needed to build against KIconThemes 6a75883159bbeb9, which
reverts the commit that adds the path to the addAppDir call. From
looking at the API docs, this *seems* to be the right fix, but I don't
know how to test this properly, so I can't confirm that this actually
works. It builds though, and that's at least a step forward.

If someone can test this, that would be very useful.

CCMAIL:mklapetek@kde.org
CCMAIL:plasma-devel@kde.org

Reviewed-by: Eike Hein on IRC