[klipper] Add ClipboardService and ClipboardJob
The ClipboardEngine provides a service which in turn starts a
ClipboardJob. The following operations are globally supported:
* configureKlipper
* clearHistory

And the following operations are supported on uuid (base64):
* select
* remove
* edit
* barcode
* action

barcode is only supported if klipper gets compiled with prison support,
this is announced as data element "supportsBarcodes" in the "clipboard"
source of the DataEngine.