Startkde: Remove KLOCALE_LANGUAGES
KLOCALE_LANGUAGES was used by the kde4 ksplash in order to to know which
language to show. This environment variable is no longer used by the qml
based ksplash. It makes no sense to have it.

Additionally, this means we can stop linking against kdelibs4support.
This is important cause kdostartupconfig blocks the rest of the boot
sequence. On my system it causes a good 0.3 - 0.4 seconds delay. By no
longer linking to kdelibs4support it takes less than 0.1 seconds and no
longer shows up in the bootchat logs.

REVIEW: 119103