Polish screen connection event
Make sure the screen is not added until the containment is defined and the
screen is shown.
Don't check if a screen is redundant to itself.
Also delay output redundancy using the timer, until the screen is taken
care for. Also delay the timer a bit further as all the screen management
takes quite some time.