Don't remove dbus-activated plasmoids too early.
The systemtray plugin for dbus-activated plasmoids removed the
plasmoid as soon as a service vanished, that matched the pattern.
This behavior broke as soon as there're multiple matchin services,
e.g. the mediacontroller plasmoid when multiple mpris2-mediaplayers
are started. As soon as you quit one application, the plasmoid would
get removed.

With this patch the plasmoid only gets removed after the last service
quit.