Polish plasmoid initialization, screenForContainment
So far, when initializing plasma shell some plasmoids would trigger a
warning in ShellCorona::screenForContainment saying that the requested
screen was not found.
This patch fixes the issue by introducing the screens to the views list
earlier, this way when screenForContainment is called, we already have
the list initialized. Now the problem is that we get some inconsistencies,
because QApp::processEvents is called from within View::setContainment.
This results in requests to the PlasmaShell on half-baked instances and
resulted in problems. That's why I introduced the loading attribute that
wraps setContainment calls.

REVIEW: 119554