Fix broken creation of kstartupconfigfiles
kstartupconfigfiles contains a list of files to compare mtimes on to see
if we need to rebuild kstartupconfig. This wasn't being created
correctly so we failed to rebuild if any configs updated.

This was broken in Qt5 porting:
-        const QStringList dirs =
KGlobal::dirs()->kfsstnd_prefixes().split( KPATH_SEPARATOR, QString::SkipEmptyParts);
+        const QStringList dirs =
QStandardPaths::standardLocations(QStandardPaths::GenericDataLocation);

REVIEW: 119669