Simplify Klipper plasmoid and implement keyboard navigation
This patch simplifies the new Klipper plasmoid by removing the custom highlight
implementation and using ListView's highlight property similar to Kickoff.
This also gets rid of the annoying lagging-behind selection rectangle and it
not properly repositioning when scrolling the list.

I also implemented keyboard navigation similar to Kickoff.
Open the plasmoid and start typing and your input will be forwarded to
the filter input. Use arrow up and down to navigate between entries and
Return to select them.
Not yet reachable via keyboard only are the action keys.

REVIEW: 119772