BalooSearchRunner: Set the relevance for each result
The Plasma::QueryMatch operator < sorts first based on relevance and
then based on the text. So even though we try and order stuff correctly,
mostly clients just sort it based on relevance (not stable sort) which
results in the results being mixed up. Baloo takes a lot of pain to
internally determine which results should be on top.