non-recursive version that doesn't march panels to the last screen
rather than implementing it as a (harder to read) recursive
function, simply start from the second-to-last item in the list and
march backwards until it hits the index being removed, moving the
panels and desktop forward as it goes.

previously it was starting from the index being removed, moving
the panels to the next screen, then processing the next screen
at which point all the panels on idx would now be on idx+1's screen
and they'd get pushed forward again. on a 4 screen system with
panels on each i bet that was fun to watch.

this should be both less buggy and easier to comprehend code.