Fix the screen locker to be able to show the user switch dialog multiple times.
The user switching dialog from the screen locker was destroyed when the
cancel button was pressed. As a result you coudn't open it again.

The screen locker uses a PlasmaComponents.PageStack to show the password input
and user switching dialogs.
When a dialog is popped from the stack it is destroyed by PageStack.

Fix this by turning userSessionsUI into a Component so it is automatically
created.

This work has been sponsored by the City of Munich's LiMux project.

Forward port of kde-workspace 3341db4430026914da4c6b5f90c1ef6920bae9d1