move WallpaperPackage into a pluging called Wallpaper/Images
this will allow it to be shared by other plugins out there. it does mean
sacrificing the generic "find best paper size" functionality which was
tightly bound to the image wallpaper (in plasma4, this was Plasma::Wallpaper,
so acceptable). Finding the best image is now left up to the plugin,
though I imagine that other plugins will want this eventually as well
and it will find its way into a library somewhere or else just be
copy-and-pasted a lot.