Don't start Passive
Systray is really broken when you have a DBus activated task that starts
passively. It adds it to some model and then moves it around and then eventually
everything will just break. This needs fixing in the systray, since there
will eventually be plasmoids which should start passively but atm the
mediacontroller is the only one. So I'm trying to make it look okay here.