Add a workaround for when autostart2 fails
If klauncher can't launch any of the services within the phase2
autoStart2Done will be never called, hence leaving ksmservice in an
invalid status.

Since at this moment is almost impossible for klauncher to make sure
that autoStart2Done is always called we have to add a work around the
same way we have done with all the other phases.

This workaround is specially useful since it is in phase2 where we
launch third party stuff.