Add a context menu action to remove applications.
This introduces an optional build-time dependency on
PackageKit-Qt.

Which package manager to start is currently handled via a
'removeApplicationCommand' config key which defaults to
Muon Discover, which will be moved to workspace soon. It's
been discussed to add a Default Applications entry for
this in the future, though; this small addition will allow
us to gather some experience with tighter package manage-
ment integration in Plasma.

Derived from original code by Aleix Pol.
REVIEW:120318