Add hint-dont-superimpose-fill
This allows the theme to decide it does not want us to just show a battery frame
and then add the individual fill pieces but to provide a different battery for
every fill. The current Breeze theme does this by providing a red Fill20 and Fill0
and to prevent superimposing a full red battery ontop of a black one, the Battery svg
can have this item which hides the battery frame.