[applets/digital-clock] Split the clock label into 3 separate ones
Right now the clock applet is using one single label (besides the
sizehelper) to fit time, timezone and also date. This has several
drawbacks like for example the whole label must use one font size, one
eliding etc.

Putting everything into its own label gives way more flexibility as we
can position and size things independently - for example making the time
bigger font than the timezone or date and elide long timezone name
without omitting the date. The sizing is now also simpler and more
robust.

This patch adds the time and timezone labels into Flow which changes
direction based on the layout (in vertical panels layouts things
vertically, horizontal panels either vertically or horizontally
depending on date being shown or not), the date label is then always
appended at the bottom. The reason for dateLabel not being in the Flow
is that in horizontal panels we want the date label always go to the
bottom and be center aligned and Flow does not support horizontal
alignment of its children. Two anchors seems much easier in this case.

This removes two i18n strings and works around having to add another one
(line 352 in the patch) - it's not kosher but I did it so it can still
be merged for 5.1.

Finally, this adds states for separate handling of vertical and
horizontal panels, which has cleaned the code quite a lot from all the
"vertical ? complex_a_stuff : complex_b_stuff".

REVIEW: 120461