Baloo Runner: Lower relevance because krunner does not know any better
The KRunner framework expects all runners to set a relevance for each
match, except that this relevance information is taken to globally sort
the results. So I can write a plugin which whose results will always be
on top. Yaye! </sarcasm>

Explicitly lower Baloo's relevance so that in general applications are
higher.