add possibility for l&f package to have layouts
if the l&f package has a default layout script, make it win over the one of the shell.
this should be used for distributions to customize.

since there can be one l&f package with multiple shells, a layout is specific for a particular shell package, so search the files as layouts/shellName-layout.js, like layouts/org.kde.plasma.shell-layout.js

REVIEW:121119
Reviewed by: Rohan Garg <rohangarg@kubuntu.org>