Fix lockscreen theme fallback
When the look and feel lockscreen theme failed to load for whatever reason,
it is supposed to fall back to a built-in theme to not lock the user out of the system.

The fallback did not work, however, and made the greeter crash with an invalid package
access, trapping you in an endless kscreenlocker_greet reboot loop.

REVIEW: 121333