[ksld] Provide custom QQmlNetworkAccessManagerFactory in greeter
The greeter installs a custom QQmlNetworkAccessManagerFactory which
creates QNetworkAccessManagers which do not have access to the
network.

Unfortunately this is not a 100 % secure way to disable network, any
plugin could change the factory again and thus enable network. This
is done for example by KDeclarative, so the greeter now first creates
KDeclarative and then sets the custom factory.

It is a sufficient solution to block all network access for pure
QtQuick without a binary package as we would find in a look'n'feel
package.

Network access needs to be blocked in the greeter as otherwise a
look'n'feel package were able to send the entered password (and other
information to the creator of the look'n'feel package.

REVIEW: 121847