Uncondtionally monitor for all KAMD service status until we have first loaded
There is an issue where as m_activityConsumer starts the
kactivitymanagerd, as KAMD is starting the serviceStatus will be not
running.

Whilst we are loading the kscreen config, the event loop runs and we
might find KAMD has started.

m_activityConsumer will change from "not running" to unknown, and might
still be unknown when the kscreen fetching is complete.

If that happens we want to continue monitoring for state changes so we
can finally load when it is up

BUG: 342431