Fix the infinite event recursion in the klipper popup.
Klipper crashes when any modifier key (with the exception of ALT) is pressed.

This is because it enters into an infinite recursion.
This happens because Klipper popup delegates all events to its input field.
The input field does not consume the event, so it is propagated to the parent widget,
which is the popup. The popup tries to send the event to the input field again.

Since klipper is now part of plasma, this bug kills the entire shell

REVIEW: 122106
BUG: 342947