Treat normal selected timezone as local if it matches the local
The real timezone that Local links to is also in the list of timezones
for which teh time can be displayed. For whatever reason, it can happen
that the real timezone is selected and not 'Local' and then it's not
treated as local anymore, making the "Show local timezone" setting not
work.

This fixes it.

BUG: 343908
REVIEW: 122473