[screenlocker] Perform KAuthorized check for org.freedesktop.ScreenSaver.Lock
The lock screen has an authorized action "lock_screen" which is honored
for the global shortcut. The authorized action might disallow the user
to lock the screen, so this should also be honored for other ways the
user might lock the screen, e.g. through the DBus interface.

Idle timeout is not verified through the KAuthorized action as this can
be restricted seperately.

CCBUG: 174100
REVIEW: 122509