Handle job error when showing job finished notification
Check for job error when job finishes and in case of error, set notification
summary to "%job_message%: Error", body to error text and clears the
actions (no open url).

It only works with jobs that actually sets the error text (KJob::setErrorText),
because KJob::error is not exposed in dataengine.

REVIEW: 122921