Set QGuiApplication::applicationDisplayName() too
Notifications dataengine checks the applicationDisplayName() against the
event title and if it same it allows to 'spam' aka it will not merge
similar notifications. But when we started to use KAboutData this broke
functionality. KAboutData will set only properties available in
QCoreApplication and hence applicationDisplayName will be different.

This fixes it by setting applicationDisplayName.

BUG: 345149
FIXED-IN: 5.2.2