Sanitise input to osd progress bar
Previous code checked the progress bar was visible and only then set the
osdValue (which is a var). This relies on inner workings of the OSD
service to set values in the right order.

This fixes a binding loop on startup and will be slightly faster as we
won't update value twice when showing OSD.

BUG: 344684