Display an instance of the grey login strip in the sddm theme on every screen
Display the grey login strip in sddm on every screen separated instead
of displaying one strip that goes over all screens.
The latter could result in a password promt or user selection that is
subdivided into two parts on two screens.
This commit displays the strip on every screen (like the background
image) so it won't be subdivided anymore.

REVIEW: 123124