[ksmserver] Do not try to start X11 window manager if Wayland env variables are set
If ksmserver is started in a session with WAYLAND_DISPLAY
and/or WAYLAND_SOCKET being exported we can assume that the Xserver is an
Xwayland server. In that case the wayland compositor is also the window
maager. Thus starting a window manager would either fail or create severe
havoc in case it succeeds in replacing the running window manager.

REVIEW: 123444