[klipper] Ensure global shortcut actions work
Some actions were set to not-visible in order to not show them in
the popup if not in standalone mode. This has the side effect of
global shortcuts no longer working as global shortcuts do not
trigger for not visible actions.

To fullfill both needs the check is now done to only add the actions
to the popup if it's in standalone mode, but the visibility is not
changed.

BUG: 345945
FIXED-IN: 5.3.1
REVIEW: 123727