Switch the login sound to Phonon directly...for now
With its current architecture, KNotification can cause crashes on logout
due to use of Q_GLOBAL_STATIC and threads and Phonon and cleanup by main
thread. So this replaces the KNotification-in-a-thread with Phonon
directly.

This is exactly what KNotification would do. This is for the time being
until the crash on logout is sorted out.

Additionally, this also fixes logout sound which was missing before.
This uses normal KNotification as at that point we don't need to be
threading or anything, so KNotification is just safe.

REVIEW: 123834