Add support to mark the PanelView as a Panel on Wayland
This change introduces the first part of Wayland integration. If
running on platform Wayland we try to bind the org_kde_plasma_shell
interface through the help of KWayland client library. If this interface
is available we can use it to create a PlasmaShellSurface for the
PanelView's window and mark that as a panel.

A compositor with support for the org_kde_plasma_shell interface can
use this information to make the window operate as a panel.

REVIEW: 124054