Replace other Notifications services when Plasma's notifications are enabled
This patch makes Plasma always be the Notification service provider if
that option is enabled in the applet settings and/or if the applet is
present somewhere (otherwise the dataengine is not loaded). On startup,
it will get the PID of the current Notifications service, send SIGTERM
to it and register its own service.

REVIEW: 124102