Fall back to averaging percentage when there are no energy values reported
According to the UPower specification UPS batteries do not suppliy energy values (Wh)
thus our calculation breaks when there's just an UPS and we start blinking because
we think the total percentage is zero. In case we do have batteries but no energy
just average the percentages. Still not perfect but close enough.

BUG: 348588
FIXED-IN: 5.4.0