[shell] Add Wayland support to PanelShadow
This change adds support for shadows using the KWin specific shadow
protocol provided by KWayland. The protocol is inspired by the X11
protocol, so it can be easily integrated. The protocol takes a buffer
for each of the shadow parts and provides the same offsets.

To make this possible the code is better split into X11 and Wayland
specific parts.

REVIEW: 124360